backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
langchain
langchain-openai
langchain-pinecone
pinecone-client
python-dotenv
python-multipart
PyMuPDF
